Set plugin release number in build workflow

+++ scripts/build-package-zip.sh
@@ -11,4 +11,17 @@ cp Pokeheim/Package/{icon.png,manifest.json} staging/
 cp README.md staging/
 cp -a Pokeheim/Assets/Translations staging/Pokeheim/Assets/
 (cd staging; zip -r9 ../Pokeheim.zip *)
+
+set +x
+manifest_version=$(cat staging/manifest.json | jq -r .version_number)
+dll_version=$(monodis --assembly staging/Pokeheim/Pokeheim.dll \
+              | grep Version | cut -f 2 -d : | tr -d ' ')
+
+if [[ "$manifest_version.0" != "$dll_version" ]]; then
+  echo "Version mismatch!"
+  echo "  Manifest version $manifest_version"
+  echo "  DLL version $dll_version"
+  exit 1
+fi
+
 rm -rf staging
